// mollys-movies-ui/stubs/api/model/movie.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { LocalMovieSource } from './local-movie-source';
import { MovieDownload } from './movie-download';
import { Torrent } from './torrent';

/**
 * 
 * @export
 * @interface Movie
 */
export interface Movie {
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    imdbCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    language?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    year?: number;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    rating?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    youTubeTrailerCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    imageFilename?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof Movie
     */
    genres?: Array<string> | null;
    /**
     * 
     * @type {Array<Torrent>}
     * @memberof Movie
     */
    torrents?: Array<Torrent> | null;
    /**
     * 
     * @type {LocalMovieSource}
     * @memberof Movie
     */
    localSource?: LocalMovieSource;
    /**
     * 
     * @type {MovieDownload}
     * @memberof Movie
     */
    download?: MovieDownload;
}